HTML Output for White Bread (#81)
* Fetch outputer from application configuration

We configure the current console outputer as a default under
application/0 in the Mix file, fetch the module name, then call/cast
into it.

* Test that the default outputer remains to the console

* Introduce HTML outputer which accumulates report state

* Format module documentation and improve variable names

* Introduce casts for remaing report constructs

* Narrow down on arguments in function heads and guards

* Build HTML document when server is stopped

Introduces a module for building HTML. Print to the console for the
time being.

* Write out HTML report to a file

Hard coded to write to ~/report.html for the time being.

* Make report path configurable

* Document HTML output configuration in README

* Improve HTML report with a descriptive line

* Write directory tree for the HTML document's path if it doesn't exist

* Make changes to HTML outputter following review

Thank you to Szymon Mentel for suggested improvements. This introduces
the PathError exception too.

* Follow style convenetions and update the Credo dependency

Credo wouldn't work as it used an internal Elixir API call that was
changed.